SDL Anonymizer Code review changes: 1. Renamed folder for styles 2. Moved Studio Control and plugin view in a separate fodler
