+++ SDLBatchAnonymize/SDLBatchAnonymize/Control/BatchAnonymizerHostControl.cs
@@ -1,10 +1,9 @@
 ﻿using System.Windows.Forms;
 using Sdl.Community.SDLBatchAnonymize.BatchTask;
-using Sdl.Community.SDLBatchAnonymize.Service;
 using Sdl.Community.SDLBatchAnonymize.ViewModel;
 using Sdl.Desktop.IntegrationApi;
 
-namespace Sdl.Community.SDLBatchAnonymize.Ui
+namespace Sdl.Community.SDLBatchAnonymize.Control
 {
 	public partial class BatchAnonymizerHostControl : UserControl, ISettingsAware<BatchAnonymizerSettings>
 	{
@@ -12,7 +11,7 @@ namespace Sdl.Community.SDLBatchAnonymize.Ui
 		{
 			InitializeComponent();
 			BatchAnonymizerSettingsViewModel = new BatchAnonymizerSettingsViewModel();
-			var batchAnonymizerControl = new BatchAnonymizerSettingsWpfControl
+			var batchAnonymizerControl = new View.BatchAnonymizerSettingsView
 			{
 				DataContext = BatchAnonymizerSettingsViewModel
 			};
