+++ SDLBatchAnonymize/SDLBatchAnonymize/Styles/General.xaml
