--- SDLBatchAnonymize/SDLBatchAnonymize/Ui/BatchAnonymizerSettingsWpfControl.xaml.cs
@@ -1,13 +1,0 @@
-﻿namespace Sdl.Community.SDLBatchAnonymize.Ui
-{
-	/// <summary>
-	/// Interaction logic for BatchAnonymizerSettingsWpfControl.xaml
-	/// </summary>
-	public partial class BatchAnonymizerSettingsWpfControl
-	{
-		public BatchAnonymizerSettingsWpfControl()
-		{
-			InitializeComponent();
-		}
-	}
-}
