+++ SDLBatchAnonymize/SDLBatchAnonymize/View/BatchAnonymizerSettingsView.xaml.cs
@@ -1,0 +1,13 @@
+﻿namespace Sdl.Community.SDLBatchAnonymize.View
+{
+	/// <summary>
+	/// Interaction logic for BatchAnonymizerSettingsWpfControl.xaml
+	/// </summary>
+	public partial class BatchAnonymizerSettingsView
+	{
+		public BatchAnonymizerSettingsView()
+		{
+			InitializeComponent();
+		}
+	}
+}
